requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
reportlab
python-dotenv
python-multipart
google-generativeai
anthropic
sqlalchemy
jinja2
